requirements.txt
django>=5.0
djangorestframework
django-cors-headers
